refactor: move custom objects to game_objects.py
